UART System Module
82b2d1d0bf9296761c7ae9a5bf27ba37a6b0513588e466178837c090e91cc42e

- System Modules
  - Extended kernel api to include:-
    - `vsnprintf`
    - `outb`, `outw`, `outl`, `inb`, `inw`, `inl`
    - `kmalloc` and `kfree`
    - `register_device` and `unregister_device`
  - Implemented `uart.sys`
  - Special `sysmod_header` region in linker
  - Renamed `loader.c` to `sysmods.c`
- FXTools
  - Added `QEMU_FULLSCREEN` to `statejson`
  - `make` does not continue to next step if an error occurred.

+++ arch/x86_64/memory/vmm.c
@@ -28,12 +28,12 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 #include "memory/vmm.h"
 
-const uintptr_t PAGE_PRESENT = 0x1;  // 10e0 in binary - If page is in RAM
-const uintptr_t PAGE_RW      = 0x2;  // 10e1 in binary - 0 = Read-only,   1 = Read/Write
-const uintptr_t PAGE_USER    = 0x4;  // 10e2 in binary - 0 = Kernel only, 1 = Everyone
-const uintptr_t PAGE_PWT     = 0x8;  // 10e3 in binary - Writes go to cache and memory immediately
-const uintptr_t PAGE_PCD     = 0x10; // 10e4 in binary - Completely disables CPU caching for that page
-const uintptr_t PAGE_CACHE   = 0x0;  // Not present in x86, does nothing, but required stub
+const uint64_t PAGE_PRESENT = 0x1;  // 10e0 in binary - If page is in RAM
+const uint64_t PAGE_RW      = 0x2;  // 10e1 in binary - 0 = Read-only,   1 = Read/Write
+const uint64_t PAGE_USER    = 0x4;  // 10e2 in binary - 0 = Kernel only, 1 = Everyone
+const uint64_t PAGE_PWT     = 0x8;  // 10e3 in binary - Writes go to cache and memory immediately
+const uint64_t PAGE_PCD     = 0x10; // 10e4 in binary - Completely disables CPU caching for that page
+const uint64_t PAGE_CACHE   = 0x0;  // Not present in x86, does nothing, but required stub
 
 #define PAGING_BIT  0x80000000
 #define PAGE_WP_BIT 0x00010000
